% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery.R
\name{battery}
\alias{battery}
\title{battery}
\usage{
battery(
  title = setNames(c("LongGold Test Battery", "LongGold Test Batterie"),
    psyquest::languages()),
  questionnaires,
  languages = psyquest::languages(),
  dict = psyquest::psyquest_dict,
  admin_password = "conifer",
  researcher_email = NULL,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Character scalar) Title of the umbrella test battery.}

\item{questionnaires}{(Character vector)}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{admin_password}{Password to access the admin panel.}

\item{researcher_email}{Researcher's email; used in participant help message.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto"
for default validation which means ID should consist only of alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{battery}()}.}
}
\description{
This function defines a battery, a series of questionnaires.
Use this function if you want to create a battery of questionnaires.
}
