% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFC.R
\name{BFC}
\alias{BFC}
\title{BFC}
\usage{
BFC(label = "BFC", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{BFC}()}.}
}
\description{
This function defines a Big Five Inventory (10-items, BFI-10, BFC) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the BFC in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the BFC,
consider using \code{\link{BFC_standalone}()}.
}
