% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BFC_standalone}
\alias{BFC_standalone}
\title{BFC Standalone}
\usage{
BFC_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), \code{"de"} (German) and \code{"de_d"} (formal German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the BFC questionnaire.
BFC stands for 'Big Five Inventory 10-items for Parents' (about a child).
}
