% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFT.R
\name{BFT}
\alias{BFT}
\title{BFT}
\usage{
BFT(label = "BFT", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{BFT}()}.}
}
\description{
This function defines a Big Five Inventory (10-items, BFI-10, BFT) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the BFT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the BFT,
consider using \code{\link{BFT_standalone}()}.
}
