% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{BFT_standalone}
\alias{BFT_standalone}
\title{BFT Standalone}
\usage{
BFT_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), \code{"de"} (German) and \code{"de_f"} (formal German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the BFT questionnaire.
BFT stands for 'Big Five Inventory 10-items'.
}
