% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTQ.R
\name{BTQ}
\alias{BTQ}
\title{BTQ}
\usage{
BTQ(label = "BTQ", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{BTQ}()}.}
}
\description{
This function defines a BTQ module for incorporation into a
psychTestR timeline.
Use this function if you want to include the BTQ in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the BTQ,
consider using \code{\link{BTQ_standalone}()}.
}
