% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{CCM_languages}
\alias{CCM_languages}
\title{CCM_languages
Lists the languages available for the CCM questionnaire.}
\usage{
CCM_languages()
}
\description{
CCM_languages
Lists the languages available for the CCM questionnaire.
}
