% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{CCM_standalone}
\alias{CCM_standalone}
\title{CCM Standalone
This function launches a standalone testing session for the CCM questionnaire.}
\usage{
CCM_standalone(languages = CCM_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{CCM_standalone}()}.}
}
\description{
CCM Standalone
This function launches a standalone testing session for the CCM questionnaire.
}
