% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{CMI_standalone}
\alias{CMI_standalone}
\title{CMI Standalone}
\usage{
CMI_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English),  \code{"de_f"} (German), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CMI questionnaire.
CMI stands for 'Children  Musicality Index'.
}
