% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{CMS_standalone}
\alias{CMS_standalone}
\title{CMS Standalone}
\usage{
CMS_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English),  \code{"de_f"} (German), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CMS questionnaire.
CMS stands for 'Children  Musicality Questionnaire'.
}
