% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMT.R
\name{CMT}
\alias{CMT}
\title{CMT}
\usage{
CMT(label = "CMT", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"(Re)productive competence"}, \code{"Reflection competence"}, \code{"Psychomotoric competence"}, and \code{"Social competence"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{CMT}()}.}
}
\description{
This function defines a CMT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the CMT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the CMT,
consider using \code{\link{CMT_standalone}()}.
}
