% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{CMT_languages}
\alias{CMT_languages}
\title{CMT_languages
Lists the languages available for the CMT questionnaire.}
\usage{
CMT_languages()
}
\description{
CMT_languages
Lists the languages available for the CMT questionnaire.
}
