% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{CRT_standalone}
\alias{CRT_standalone}
\title{CRT Standalone}
\usage{
CRT_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"}/\code{"de_f"} (informal/formal German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CRT questionnaire.
CRT stands for 'Cognitive Reflection Test'.
}
