% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAC.R
\name{DAC}
\alias{DAC}
\title{DAC}
\usage{
DAC(label = "DAC", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{DAC}()}.}
}
\description{
This function defines a DAC module for incorporation into a
psychTestR timeline.
Use this function if you want to include the DAC in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the DAC,
consider using \code{\link{DAC_standalone}()}.
}
