% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{DAC_languages}
\alias{DAC_languages}
\title{DAC_languages
Lists the languages available for the DAC questionnaire.}
\usage{
DAC_languages()
}
\description{
DAC_languages
Lists the languages available for the DAC questionnaire.
}
