% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{DAC_standalone}
\alias{DAC_standalone}
\title{DAC Standalone
This function launches a standalone testing session for the DAC questionnaire.}
\usage{
DAC_standalone(languages = DAC_languages(), randomise = TRUE, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{randomise}{(Boolean scalar) Defines whether questionnaire items appear in
randomised order. Defaults to TRUE}

\item{...}{Further arguments to be passed to \code{\link{DAC_standalone}()}.}
}
\description{
DAC Standalone
This function launches a standalone testing session for the DAC questionnaire.
}
