% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG}
\alias{DEG}
\title{DEG}
\usage{
DEG(
  label = "DEG",
  dict = psyquest::psyquest_dict,
  items = items,
  subscales = subscales,
  language = language,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label to give the DEG results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{items}{(Data frame) The items to be included in the questionnaire.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
When no subscales are provided all subscales are selected.}

\item{language}{Language the questionnaire is rendered in.}

\item{...}{Further arguments to be passed to \code{\link{DEG}()}.}
}
\description{
This function defines a DEG module for incorporation into a
psychTestR timeline.
Use this function if you want to include the DEG in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the DEG, consider using \code{\link{DEG_standalone}()}.
}
