% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{DEG_languages}
\alias{DEG_languages}
\title{DEG_languages
Lists the languages available for the DEG questionnaire.}
\usage{
DEG_languages()
}
\description{
DEG_languages
Lists the languages available for the DEG questionnaire.
}
