% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{DEG_standalone}
\alias{DEG_standalone}
\title{DEG Standalone
This function launches a standalone testing session for the DEG questionnaire.}
\usage{
DEG_standalone(languages = DEG_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{DEG_standalone}()}.}
}
\description{
DEG Standalone
This function launches a standalone testing session for the DEG questionnaire.
}
