% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSR.R
\name{FSR}
\alias{FSR}
\title{FSR}
\usage{
FSR(
  label = "FSR",
  dict = psyquest::psyquest_dict,
  subscales = c(),
  short_version = F,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Absorption"}, \code{"Fluency of performance"}, \code{"Demands"},
\code{"Skills"}, \code{"Demand Fit"}
and \code{"Importance"}.
If no subscales are provided all subscales are selected.}

\item{short_version}{(boolean) Use short version of test}

\item{...}{Further arguments to be passed to \code{\link{main_test}()}.}
}
\description{
This function defines a Flow State Scale (Marsh & Jackson, 1996) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the FSR in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the FSR,
consider using \code{\link{FSR_standalone}()}.
}
