% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{FSR_standalone}
\alias{FSR_standalone}
\title{FSR Standalone}
\usage{
FSR_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Absorption"}, \code{"Fluency of performance"}, \code{"Demands"},
\code{"Skills"}, \code{"Demand Fit"}
and \code{"Importance"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{FSR}()}.}
}
\description{
This function launches a standalone testing session for the FSS questionnaire.
FSR stands for 'Theory of Intelligence'.
}
