% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSS.R
\name{FSS}
\alias{FSS}
\title{FSS}
\usage{
FSS(
  label = "FSS",
  dict = psyquest::psyquest_dict,
  subscales = c(),
  short_version = F,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Challenge-skill balance"}, \code{"Action-awareness merging"},
\code{"Clear goals"}, \code{"Unambiguous feedback"}, \code{"Concentration on task at hand"},
\code{"Paradox of control"}, \code{"Loss of self-consciousness"}, \code{"Transformation of time"},
and \code{"Autotelic experience"}.
If no subscales are provided all subscales are selected.}

\item{short_version}{(boolean) Use short version of test}

\item{...}{Further arguments to be passed to \code{\link{main_test}()}.}
}
\description{
This function defines a Flow State Scale (Marsh & Jackson, 1996) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the FSS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the FSS,
consider using \code{\link{FSS_standalone}()}.
}
