% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{FSS_standalone}
\alias{FSS_standalone}
\title{FSS Standalone}
\usage{
FSS_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Challenge-skill balance"}, \code{"Action-awareness merging"},
\code{"Clear goals"}, \code{"Unambiguous feedback"}, \code{"Concentration on task at hand"},
\code{"Paradox of control"}, \code{"Loss of self-consciousness"}, \code{"Transformation of time"},
and \code{"Autotelic experience"}.}

\item{...}{Further arguments to be passed to \code{\link{FSS}()}.}
}
\description{
This function launches a standalone testing session for the FSS questionnaire.
FSS stands for 'Theory of Intelligence'.
}
