% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDS.R
\name{GDS}
\alias{GDS}
\title{GDS}
\usage{
GDS(label = "GDS", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) There are 4 subscales, one general subscale
and 6 items of 'dance experience observations' to be included in the questionnaire.
Possible subscales are \code{"Body Awareness"}, \code{"Social Dancing"},
\code{"Urge to Dance"}, \code{"Dance Training"}, \code{"General"},
and \code{"Observational Dance Experience"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{GDS}()}.}
}
\description{
This function defines a GDS module for incorporation into a
psychTestR timeline.
Use this function if you want to include the GDS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the GDS,
consider using \code{\link{GDS_standalone}()}.
}
