% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_item_choices}
\alias{get_item_choices}
\title{get_item_choices}
\usage{
get_item_choices(questionnaire_id, item_id, language = "en")
}
\arguments{
\item{questionnaire_id}{(three letter string) Questionnaire ID.}

\item{item_id}{(integer or character) number of item in a questionnaire}

\item{language}{(character)  language of item texts}
}
\description{
Retrieves choice labels for an items from questionnaire identified by questionnaire_id and subscale
}
