% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_item_info}
\alias{get_item_info}
\title{get_item_info}
\usage{
get_item_info(questionnaire_id, subscales, language = "en")
}
\arguments{
\item{subscales}{(character vector)  of subscale names}

\item{language}{(character)  language of item texts}

\item{questionaired_id}{(three letter string) Questionnaire ID.}
}
\description{
Retrieves for info for items from questionnaire identified by questionnaire_id and subscale
}
