% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subscales}
\alias{get_subscales}
\title{get_subscales}
\usage{
get_subscales(questionnaire_id)
}
\arguments{
\item{questionaired_id}{(three letter string) Questionnaire ID.}
}
\description{
Retrieves available subscales for a questionnaired identified by questionaired_id
}
