% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMS.R
\name{GMS}
\alias{GMS}
\title{GMS}
\usage{
GMS(
  label = "GMS",
  dict = psyquest::psyquest_dict,
  subscales = c(),
  short_version = FALSE,
  configuration_filepath = NULL,
  feedback_page = NULL,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Abilities"}, \code{"Absolute Pitch"}, \code{"Active Engagement"}, \code{"Emotions"}, \code{"General"}, \code{"Instrument"}, \code{"Musical Training"}, \code{"Perceptual Abilities"}, \code{"Singing Abilities"}, and \code{"Start Age"}.
If no subscales are provided all subscales for the questionnaire are selected.
Overrides the \code{"short_version"} argument.
Overridden by the \code{configuration_filepath} argument.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.
Overridden by the \code{configuration_filepath} and \code{"subscales"} arguments.}

\item{configuration_filepath}{(Character scalar) Optional path to a configuration file exported from the \href{https://shiny.gold-msi.org/gmsiconfigurator}{GMSI-Configurator}.
Overrides the \code{short_version} and \code{subscales} arguments.}

\item{feedback_page}{(Function) Defines a feedback page function for displaying
the results to the participant at the end of the test. Defaults to NULL.}

\item{...}{Further arguments to be passed to \code{\link{GMS}()}.}
}
\description{
This function defines a GMS module for incorporation into a
psychTestR timeline.
Use this function if you want to include the GMS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the GMS,
consider using \code{\link{GMS_standalone}()}.
}
