% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{GMS_languages}
\alias{GMS_languages}
\title{GMS_languages
Lists the languages available for the GMS questionnaire.}
\usage{
GMS_languages()
}
\description{
GMS_languages
Lists the languages available for the GMS questionnaire.
}
