% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{GMS_standalone}
\alias{GMS_standalone}
\title{GMS Standalone
This function launches a standalone testing session for the GMS questionnaire.}
\usage{
GMS_standalone(
  languages = GMS_languages(),
  subscales = NULL,
  short_version = FALSE,
  configuration_filepath = NULL,
  ...
)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}), and German (\code{"de"}).
The first language is selected by default.}

\item{subscales}{(Character vector)
Determines the subscales to be included.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE. Is overridden by the \code{configuration_filepath} argument.}

\item{configuration_filepath}{(Character scalar) Optional path to a configuration file exported from the \href{https://shiny.gold-msi.org/gmsiconfigurator}{GMSI-Configurator}. Overrides the \code{short_version} argument.}

\item{...}{Further arguments to be passed to \code{\link{GMS_standalone}()}.}
}
\description{
GMS Standalone
This function launches a standalone testing session for the GMS questionnaire.
}
