% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRT.R
\name{GRT}
\alias{GRT}
\title{GRT}
\usage{
GRT(label = "GRT", dict = psyquest::psyquest_dict, items = items, ...)
}
\arguments{
\item{label}{(Character scalar) Label to give the GRT results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{items}{(Data frame) The items to be included in the questionnaire.}

\item{...}{Further arguments to be passed to \code{\link{GRT}()}.}
}
\description{
This function defines a GRT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the GRT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the GRT,
consider using \code{\link{GRT_standalone}()}.
}
