% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{GRT_languages}
\alias{GRT_languages}
\title{GRT_languages
Lists the languages available for the GRT questionnaire.}
\usage{
GRT_languages()
}
\description{
GRT_languages
Lists the languages available for the GRT questionnaire.
}
