% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{HOP_languages}
\alias{HOP_languages}
\title{HOP_languages
Lists the languages available for the HOP questionnaire.}
\usage{
HOP_languages()
}
\description{
HOP_languages
Lists the languages available for the HOP questionnaire.
}
