% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{HUM_standalone}
\alias{HUM_standalone}
\title{HUM Standalone}
\usage{
HUM_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"it"} (Italian).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the HUM questionnaire.
HUM stands for 'Healthy-Unhealthy Music Scale'.
}
