% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{IBQ_standalone}
\alias{IBQ_standalone}
\title{IBQ Standalone}
\usage{
IBQ_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), \code{"de"} (German), \code{"de_f"} (German (formal), and \code{"ar"} (arabic).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CBQ questionnaire.
IBQ stands for 'Children's Behavior Questionnaire'.
}
