% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMI.R
\name{IMI}
\alias{IMI}
\title{IMI}
\usage{
IMI(label = "IMI", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Negative Valence"}, \code{"Help"},
\code{"Movement"}, \code{"Personal Reflection"}, \code{"Earworm Frequency"},
\code{"Earworm Duration"}, and \code{"Earworm Length"}, and \code{"Earworm Removal"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{IMI}()}.}
}
\description{
This function defines a IMI module for incorporation into a
psychTestR timeline.
Use this function if you want to include the IMI in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the IMI,
consider using \code{\link{IMI_standalone}()}.
}
