% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIC.R
\name{JIC}
\alias{JIC}
\title{JIC}
\usage{
JIC(label = "JIC", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"General"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{JIC}()}.}
}
\description{
This function defines a JIC module for incorporation into a
psychTestR timeline.
Use this function if you want to include the JIC in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the JIC,
consider using \code{\link{JIC_standalone}()}.
}
