% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{JIC_standalone}
\alias{JIC_standalone}
\title{JIC Standalone}
\usage{
JIC_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"General"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the JIC questionnaire.
JIC stands for 'Theory of Intelligence'.
}
