% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIW.R
\name{JIW}
\alias{JIW}
\title{JIW}
\usage{
JIW(label = "JIW", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"General"}, \code{"Talent"}, \code{"Enjoyment"}, \code{"Competition"}, \code{"Confidence"}
 and \code{"Growth"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{JIW}()}.}
}
\description{
This function defines a JIW module for incorporation into a
psychTestR timeline.
Use this function if you want to include the JIW in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the JIW,
consider using \code{\link{JIW_standalone}()}.
}
