% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{labelled_text_input_page}
\alias{labelled_text_input_page}
\title{Labelled text input page}
\usage{
labelled_text_input_page(
  label,
  prompt,
  one_line = TRUE,
  save_answer = TRUE,
  placeholder = NULL,
  button_text = "Next",
  width = "300px",
  height = "100px",
  validate = NULL,
  input_label = NULL,
  on_complete = NULL,
  admin_ui = NULL
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to display (character scalar or Shiny tag object).}

\item{one_line}{Whether the answer box only has one line of text.}

\item{placeholder}{Placeholder text for the text box (character scalar).}

\item{button_text}{Text for the submit button (character scalar).}

\item{width}{Width of the text box (character scalar, should be valid HTML).}

\item{height}{Height of the text box (character scalar, should be valid HTML).}
}
\description{
Creates a page where the participant puts their
answer in a text box. Derived from the psychTestR version  to include labels
}
