% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{make_ui_NAFC_radiobuttons}
\alias{make_ui_NAFC_radiobuttons}
\title{Make NAFC radiobuttons}
\usage{
make_ui_NAFC_radiobuttons(
  label,
  choiceNames,
  choiceValues,
  sublabel,
  labels = NULL,
  hide = FALSE,
  arrange_vertically = length(choiceNames) > 2L,
  id = "response_ui"
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{choiceNames}{Character vector of choice names used as labels for the radiobuttons}

\item{choiceValues}{Character vector of choice names for the radiobuttons}

\item{sublabel}{Additional text appearing above the radiobuttons (character scalar).}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{hide}{Whether the checkboxes should be hidden
(possibly to be shown later).}

\item{arrange_vertically}{Whether to arrange the checkboxes vertically
(the default) as opposed to horizontally.}

\item{id}{HTML ID for the div containing the checkboxes.}
}
\description{
Creates HTML code for n-option multiple choice response options.
}
