% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{make_ui_radiobutton_NAFC}
\alias{make_ui_radiobutton_NAFC}
\title{Make NAFC radiobuttons}
\usage{
make_ui_radiobutton_NAFC(
  label,
  choiceNames,
  choiceValues,
  subprompt,
  labels = NULL,
  hide = FALSE,
  id = "response_ui"
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{choiceNames}{Character vector of choice names used as labels for the radiobuttons.}

\item{choiceValues}{Character vector of choice names for the radiobuttons.}

\item{subprompt}{Additional text appearing above the radiobuttons (character scalar).}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{hide}{Whether the radiobuttons should be hidden (possibly to be shown later).}

\item{id}{HTML ID for the div containing the radiobuttons.}
}
\description{
Creates HTML code for n-alternative forced-choice response radiobutton options.
}
