% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.R
\name{MDS}
\alias{MDS}
\title{MDS}
\usage{
MDS(
  label = "MDS",
  target = c(en = "a music", de = "einer Musik", de_f = "einer Musik"),
  dict = psyquest::psyquest_dict,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{target}{(Character scalar or named character vector) The main target (style, artist...) for the questionnaire, if different values for different lanugages are needed then it should be a named vector with the names according to language columns in the dict.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{MDS}()}.}
}
\description{
This function defines a Music Dislikes Scale (MDS) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MDS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the MDS,
consider using \code{\link{MDS_standalone}()}.
}
