% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MDS_standalone}
\alias{MDS_standalone}
\title{MDS Standalone}
\usage{
MDS_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the HUM questionnaire.
MDS stands for 'Music Dislikes Scale'.
}
