% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MES.R
\name{MES}
\alias{MES}
\title{MES}
\usage{
MES(label = "MES", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"General"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{MES}()}.}
}
\description{
This function defines a module for incorporation the Music-Empathizing-Music-Systemizing-Inventory
(Linnemann, Kreutz, Gollwittzer, & Nater, 2018) into a
psychTestR timeline.
Use this function if you want to include the MES in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the MES,
consider using \code{\link{MES_standalone}()}.
}
