% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{MHE_languages}
\alias{MHE_languages}
\title{MHE_languages
Lists the languages available for the MHE questionnaire.}
\usage{
MHE_languages()
}
\description{
MHE_languages
Lists the languages available for the MHE questionnaire.
}
