% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MHE_standalone}
\alias{MHE_standalone}
\title{MHE Standalone
This function launches a standalone testing session for the MHE questionnaire.}
\usage{
MHE_standalone(languages = MHE_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{MHE_standalone}()}.}
}
\description{
MHE Standalone
This function launches a standalone testing session for the MHE questionnaire.
}
