% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{month_and_year_select_page}
\alias{month_and_year_select_page}
\title{New month and year select page}
\usage{
month_and_year_select_page(
  label,
  prompt,
  save_answer = TRUE,
  min_year = 1930,
  max_year = 2013,
  show_month = TRUE,
  validate = month_and_year_select_page.validate(show_month),
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL,
  failed_validation_message = psychTestR::i18n(ifelse(show_month,
    "SELECT_MONTH_AND_YEAR", "SELECT_YEAR"))
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{save_answer}{Whether or not to save the answer.}

\item{validate}{Validation function to execute.}

\item{hide_response_ui}{Whether to begin with the response interface hidden
(it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{HTML ID for the response user interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{failed_validation_message}{Message to be displayed when validation fails.}
}
\description{
Creates month and year select page.
}
