% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA.R
\name{MPA}
\alias{MPA}
\title{MPA}
\usage{
MPA(label = "MPA", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{MPA}()}.}
}
\description{
This function defines a Music Performance Anxiety Scale (MPA) module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MPA in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the MPA,
consider using \code{\link{MPA_standalone}()}.
}
