% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MPA_standalone}
\alias{MPA_standalone}
\title{MPA Standalone}
\usage{
MPA_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), \code{"lv"} (Latvian), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the MPA questionnaire.
MPA stands for 'Music Performance Anxiety Test'.
}
