% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUS.R
\name{MUS}
\alias{MUS}
\title{MUS(MUSIC)}
\usage{
MUS(label = "MUS", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
There is a set of posible subscales (MUSIC): \code{"Mellow"}, \code{"Unpretentious"}, \code{"Sophisticated"},
\code{"Intense"}, and \code{"Contemporary"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{MUS}()}.}
}
\description{
This function defines a MUS module for incorporation into a
psychTestR timeline.
Use this function if you want to include the Musical Preferences: Five factor
model in a battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the MUS,
consider using \code{\link{MUS_standalone}()}.
}
