% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MUS_standalone}
\alias{MUS_standalone}
\title{MUS Standalone}
\usage{
MUS_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
There is a set of posible subscales (MUSIC): \code{"Mellow"}, \code{"Unpretentious"}, \code{"Sophisticated"},
\code{"Intense"}, and \code{"Contemporary"}.

If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the MUSIC questionnaire.
MUS stands for 'Musical Preferences: A five-factor model (MUSIC)'.
}
