% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{NAFC_radiobuttons_page}
\alias{NAFC_radiobuttons_page}
\title{New NAFC radiobuttons page}
\usage{
NAFC_radiobuttons_page(
  label,
  prompt,
  sublabel,
  choiceNames,
  choiceValues,
  labels = NULL,
  save_answer = TRUE,
  arrange_vertically = length(choiceNames) > 2L,
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{arrange_vertically}{Whether to arrange the checkboxes vertically
(the default) as opposed to horizontally.}

\item{hide_response_ui}{Whether to begin with the response interface hidden
(it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{HTML ID for the response user interface.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for checkbox IDs
and for checkbox labels.
If named, then values will be used for checkbox IDs and names
will be used for checkbox labels.}
}
\description{
Creates an n-option multiple choice page.
}
