% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{NOMC_page}
\alias{NOMC_page}
\title{New NOMC page}
\usage{
NOMC_page(
  label,
  prompt,
  sublabel,
  choiceNames,
  choiceValues,
  labels = NULL,
  save_answer = TRUE,
  arrange_vertically = length(choiceNames) > 2L,
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL,
  force_answer = FALSE,
  failed_validation_message = "Please select at least one answer."
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{sublabel}{Additional text appearing above the checkboxes (character scalar).}

\item{choiceNames}{Character vector of choice names used as labels for the checkboxes.}

\item{choiceValues}{Character vector of choice names for the checkboxes}

\item{labels}{Optional vector of labels for the NOMC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{save_answer}{Whether or not to save the answer.}

\item{arrange_vertically}{Whether to arrange the checkboxes vertically
(the default) as opposed to horizontally.}

\item{hide_response_ui}{Whether to begin with the response interface hidden
(it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{HTML ID for the response user interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{force_answer}{Require at least one checkbox to be ticked.}

\item{failed_validation_message}{Text to display when validation fails (character scalar).}
}
\description{
Creates an n-option multiple choice page.
}
