% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{PAC}
\alias{PAC}
\title{PAC}
\usage{
PAC(label = "PAC", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Label to give the PAC results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{PAC}()}.}
}
\description{
This function defines a PAC module for incorporation into a
psychTestR timeline.
Use this function if you want to include the PAC in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the PAC,
consider using \code{\link{PAC_standalone}()}.
}
