% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{PAC_languages}
\alias{PAC_languages}
\title{PAC_languages
Lists the languages available for the PAC questionnaire.}
\usage{
PAC_languages()
}
\description{
PAC_languages
Lists the languages available for the PAC questionnaire.
}
