% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHT.R
\name{PHT}
\alias{PHT}
\title{PHT}
\usage{
PHT(
  label = "PHT",
  subscales = c(),
  short_version = FALSE,
  randomize = FALSE,
  dict = psyquest::psyquest_dict,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{PHT}()}.}
}
\description{
This function defines a PHT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the PHT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the PHT,
consider using \code{\link{PHT_standalone}()}.
}
