% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{PHT_standalone}
\alias{PHT_standalone}
\title{PHT Standalone}
\usage{
PHT_standalone(
  languages = psyquest::languages(),
  subscales = c(),
  short_version = FALSE,
  randomize = FALSE,
  ...
)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CBQ questionnaire.
PHT stands for 'Musical Phenotype'.
}
