% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{radiobutton_NAFC_page}
\alias{radiobutton_NAFC_page}
\title{New radiobutton NAFC page}
\usage{
radiobutton_NAFC_page(
  label,
  prompt,
  subprompt,
  choiceNames,
  choiceValues,
  labels = NULL,
  save_answer = TRUE,
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL,
  failed_validation_message = "Answer missing!"
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?")
or an object of class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{subprompt}{(Character scalar) Additional text in bold letters below the prompt.}

\item{choiceNames}{Character vector of choice names used as labels for the radiobuttons}

\item{choiceValues}{Character vector of choice names for the radiobuttons}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choiceNames}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{save_answer}{Whether or not to save the answer.}

\item{hide_response_ui}{Whether to begin with the response interface hidden
(it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{HTML ID for the response user interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{failed_validation_message}{(Character scalar) Text to be displayed when validation fails.}
}
\description{
Creates a radiobutton n-alternative forced choice page.
}
