% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA}
\alias{SCA}
\title{SCA}
\usage{
SCA(
  label = "SCA",
  dict = psyquest::psyquest_dict,
  subscales = c(),
  short_version = FALSE,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
When no subscales are provided all subscales are selected.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.}

\item{...}{Further arguments to be passed to \code{\link{SCA}()}.}
}
\description{
This function defines a SCA module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SCA in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the SCA,
consider using \code{\link{SCA_standalone}()}.
}
