% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SCA_languages}
\alias{SCA_languages}
\title{SCA_languages
Lists the languages available for the SCA questionnaire.}
\usage{
SCA_languages()
}
\description{
SCA_languages
Lists the languages available for the SCA questionnaire.
}
