% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCS.R
\name{SCS}
\alias{SCS}
\title{SCS}
\usage{
SCS(label = "SCS", dict = psyquest::psyquest_dict, short_version = FALSE, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.}

\item{...}{Further arguments to be passed to \code{\link{SCS}()}.}
}
\description{
This function defines a SCS module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SCS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the SCS,
consider using \code{\link{SCS_standalone}()}.
}
