% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SCS_languages}
\alias{SCS_languages}
\title{SCS_languages
Lists the languages available for the SCS questionnaire.}
\usage{
SCS_languages()
}
\description{
SCS_languages
Lists the languages available for the SCS questionnaire.
}
