% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SCS_standalone}
\alias{SCS_standalone}
\title{SCS Standalone
This function launches a standalone testing session for the SCS questionnaire.}
\usage{
SCS_standalone(languages = SCS_languages(), short_version = FALSE, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.}

\item{...}{Further arguments to be passed to \code{\link{SCS_standalone}()}.}
}
\description{
SCS Standalone
This function launches a standalone testing session for the SCS questionnaire.
}
