% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDQ.R
\name{SDQ}
\alias{SDQ}
\title{SDQ}
\usage{
SDQ(label = "SDQ", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
When no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{SDQ}()}.}
}
\description{
This function defines a SDQ module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SDQ in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the SDQ,
consider using \code{\link{SDQ_standalone}()}.
}
