% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SDQ_languages}
\alias{SDQ_languages}
\title{SDQ_languages
Lists the languages available for the SDQ questionnaire.}
\usage{
SDQ_languages()
}
\description{
SDQ_languages
Lists the languages available for the SDQ questionnaire.
}
