% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SEM_languages}
\alias{SEM_languages}
\title{SEM_languages
Lists the languages available for the SEM questionnaire.}
\usage{
SEM_languages()
}
\description{
SEM_languages
Lists the languages available for the SEM questionnaire.
}
