% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SES.R
\name{SES}
\alias{SES}
\title{SES}
\usage{
SES(label = "SES", dict = psyquest::psyquest_dict)
}
\arguments{
\item{label}{(Character scalar) Label to give the SES results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}
}
\description{
This function defines a SES module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SES in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the SES,
consider using \code{\link{SES_standalone}()}.
}
