% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SES_languages}
\alias{SES_languages}
\title{SES_languages
Lists the languages available for the SES questionnaire.}
\usage{
SES_languages()
}
\description{
SES_languages
Lists the languages available for the SES questionnaire.
}
