% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SES_standalone}
\alias{SES_standalone}
\title{SES Standalone
This function launches a standalone testing session for the SES questionnaire.}
\usage{
SES_standalone(languages = SES_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{SES_standalone}()}.}
}
\description{
SES Standalone
This function launches a standalone testing session for the SES questionnaire.
}
