% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{SOS_languages}
\alias{SOS_languages}
\title{SOS_languages
Lists the languages available for the SOS questionnaire.}
\usage{
SOS_languages()
}
\description{
SOS_languages
Lists the languages available for the SOS questionnaire.
}
