% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SOS_standalone}
\alias{SOS_standalone}
\title{SOS Standalone
This function launches a standalone testing session for the SOS questionnaire.}
\usage{
SOS_standalone(languages = SOS_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{SOS_standalone}()}.}
}
\description{
SOS Standalone
This function launches a standalone testing session for the SOS questionnaire.
}
