% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWB.R
\name{SWB}
\alias{SWB}
\title{SWB}
\usage{
SWB(label = "SWB", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{SWB}()}.}
}
\description{
This function defines a SWB module (Subjective Wellbeing Scale) for incorporation into a
psychTestR timeline.
Use this function if you want to include the SWB in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standaSWBe implementation of the SWB,
consider using \code{\link{SWB_standaSWBe}()}.
}
