% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SWL_standalone}
\alias{SWL_standalone}
\title{SWL Standalone}
\usage{
SWL_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), \code{"de"} (German), \code{"de_f"} (Formal German) and \code{"lv"} (Latvian).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the HOP questionnaire.
SWL stands for 'Satisfaction with Life Scale'.
}
