% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOI.R
\name{TOI}
\alias{TOI}
\title{TOI}
\usage{
TOI(label = "TOI", dict = psyquest::psyquest_dict)
}
\arguments{
\item{label}{(Character scalar) Label to give the TOI results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}
}
\description{
This function defines a TOI module for incorporation into a
psychTestR timeline.
Use this function if you want to include the TOI in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the TOI, consider using \code{\link{TOI_demo}()}.
For a standalone implementation of the TOI,
consider using \code{\link{TOI_standalone}()}.
}
