% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{TOI_languages}
\alias{TOI_languages}
\title{TOI_languages
Lists the languages available for the TOI questionnaire.}
\usage{
TOI_languages()
}
\description{
TOI_languages
Lists the languages available for the TOI questionnaire.
}
