% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{TOM_languages}
\alias{TOM_languages}
\title{TOM_languages
Lists the languages available for the TOM questionnaire.}
\usage{
TOM_languages()
}
\description{
TOM_languages
Lists the languages available for the TOM questionnaire.
}
