% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{TOM_standalone}
\alias{TOM_standalone}
\title{TOM Standalone
This function launches a standalone testing session for the TOM questionnaire.}
\usage{
TOM_standalone(languages = TOM_languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}), and German (\code{"de"}).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Entity"}, \code{"Gift"}, \code{"Improvement"}, \code{"Incremental"}, \code{"Learning"}, and \code{"Stable"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{TOM_standalone}()}.}
}
\description{
TOM Standalone
This function launches a standalone testing session for the TOM questionnaire.
}
