% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{TPI_languages}
\alias{TPI_languages}
\title{TPI_languages
Lists the languages available for the TPI questionnaire.}
\usage{
TPI_languages()
}
\description{
TPI_languages
Lists the languages available for the TPI questionnaire.
}
