% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{TPI_standalone}
\alias{TPI_standalone}
\title{TPI Standalone
This function launches a standalone testing session for the TPI questionnaire.}
\usage{
TPI_standalone(languages = TPI_languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{subscales}{(Character vector)
Determines the subscales to be included.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{TPI_standalone}()}.}
}
\description{
TPI Standalone
This function launches a standalone testing session for the TPI questionnaire.
}
