% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery.R
\name{battery}
\alias{battery}
\title{battery}
\usage{
battery(
  title = "LongGold Test Umbrella Test",
  documentation = "LGTUT",
  admin_password = "conifer",
  researcher_email = "musicsophistication@gmail.com",
  languages = c("EN", "DE"),
  validate_id = "auto",
  dict = psyquest::psyquest_dict,
  questionnaires = questionnaires,
  ...
)
}
\arguments{
\item{dict}{The psyquest dictionary used for internationalisation.}

\item{label}{(Character scalar) Label to give the DAC results in the output file.}
}
\description{
This function defines a battery, a series of questionnaires.
Use this function if you want to create a battery of tests.
}
