% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{CCM_standalone}
\alias{CCM_standalone}
\title{CCM Standalone}
\usage{
CCM_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"General"}, and \code{"Extra"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the CCM questionnaire.
CCM stands for 'Concurrent Musical Activities'.
}
