% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG}
\alias{DEG}
\title{DEG}
\usage{
DEG(
  label = "DEG",
  dict = psyquest::psyquest_dict,
  subscales = c(),
  language = "en",
  year_range = c(1930, 2013),
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) (i18n_dict) (i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are "Best Shot", "Hearing Impairment", "Type of Hearing Impairment", "Gender", "Age", "Nationality", "Country Formative Years", "First Language", "Second Language", and "Handedness".
If no subscales are provided all subscales are selected.}

\item{language}{Language the questionnaire is rendered in.}

\item{year_range}{(length 2 int vector) The minimum and maximum year for birth date dropdown boxes.}

\item{...}{Further arguments to be passed to \code{\link{DEG}()} (e.g., can be used to set choice of country of residence.)}
}
\description{
This function defines a DEG module for incorporation into a
psychTestR timeline.
Use this function if you want to include the DEG in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the DEG, consider using \code{\link{DEG_standalone}()}.
}
