% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_item_info}
\alias{get_item_info}
\title{get_item_info}
\usage{
get_item_info(questionnaire_id, subscales = c(), language = "en")
}
\arguments{
\item{questionnaire_id}{(three letter string) Questionnaire ID.}

\item{subscales}{(character vector)  of subscale names}

\item{language}{(character)  language of item texts}
}
\description{
Retrieves for info for items from questionnaire identified by questionnaire_id and subscale
}
