% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_items}
\alias{get_items}
\title{get_items}
\usage{
get_items(
  q_id,
  subscales = c(),
  short_version = FALSE,
  configuration_filepath = NULL
)
}
\arguments{
\item{subscales}{(character vector)  of subscale names}

\item{short_version}{(logical)  Short version (for some tests)}

\item{label}{(three letter string) Questionnaire ID.}

\item{configuationr_filepath}{(string)  Config file for GMS}
}
\description{
Retrieves items for a questionnaire identified by label and subscales
}
