% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMS.R
\name{GMS}
\alias{GMS}
\title{GMS}
\usage{
GMS(label = "GMS", dict = psyquest::psyquest_dict, items = items, ...)
}
\arguments{
\item{label}{(Character scalar) Label to give the GMS results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{items}{(Data frame) The items to be included in the questionnaire.}

\item{...}{Further arguments to be passed to \code{\link{GMS}()}.}
}
\description{
This function defines a GMS module for incorporation into a
psychTestR timeline.
Use this function if you want to include the GMS in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the GMS,
consider using \code{\link{GMS_standalone}()}.
}
