% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{GMS_standalone}
\alias{GMS_standalone}
\title{GMS Standalone
This function launches a standalone testing session for the GMS questionnaire.}
\usage{
GMS_standalone(languages = DAC_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{DAC_standalone}()}.}
}
\description{
GMS Standalone
This function launches a standalone testing session for the GMS questionnaire.
}
