% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{GMS_standalone}
\alias{GMS_standalone}
\title{GMS Standalone}
\usage{
GMS_standalone(
  languages = psyquest::languages(),
  subscales = NULL,
  short_version = FALSE,
  configuration_filepath = NULL,
  ...
)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Abilities"}, \code{"Absolute Pitch"}, \code{"Active Engagement"}, \code{"Emotions"}, \code{"General"}, \code{"Instrument"}, \code{"Musical Training"}, \code{"Perceptual Abilities"}, \code{"Singing Abilities"}, and \code{"Start Age"}.
If no subscales are provided all subscales for the questionnaire are selected.
Overrides the \code{"short_version"} argument.
Overridden by the \code{configuration_filepath} argument.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.
Overridden by the \code{configuration_filepath} and \code{"subscales"} arguments.}

\item{configuration_filepath}{(Character scalar) Optional path to a configuration file exported from the \href{https://shiny.gold-msi.org/gmsiconfigurator}{GMSI-Configurator}.
Overrides the \code{short_version} and \code{subscales} arguments.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the GMS questionnaire.
GMS stands for 'Goldsmiths Musical Sophistication Index'.
}
