% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRT.R
\name{GRT}
\alias{GRT}
\title{GRT}
\usage{
GRT(label = "GRT", dict = psyquest::psyquest_dict, feedback_page = NULL, ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{feedback_page}{(Function) Defines a feedback page function for displaying
the results to the participant at the end of the test. Defaults to NULL.}

\item{...}{Further arguments to be passed to \code{\link{GRT}()}.}
}
\description{
This function defines a GRT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the GRT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the GRT,
consider using \code{\link{GRT_standalone}()}.
}
