% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{HOP_standalone}
\alias{HOP_standalone}
\title{HOP Standalone}
\usage{
HOP_standalone(languages = HOP_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{HOP_standalone}()}.}
}
\description{
This function launches a standalone testing session for the HOP questionnaire.
}
