% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{make_ui_month_and_year_select}
\alias{make_ui_month_and_year_select}
\title{Make month and year selectboxes}
\usage{
make_ui_month_and_year_select(
  label,
  hide = FALSE,
  id = "response_ui",
  min_year = 1930,
  max_year = 2013,
  show_month = TRUE
)
}
\arguments{
\item{label}{Label for the current page (character scalar).}

\item{hide}{Whether the checkboxes should be hidden
(possibly to be shown later).}

\item{id}{HTML ID for the div containing the checkboxes.}
}
\description{
Creates HTML code for month and year selectboxes.
}
