% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{make_ui_NOMC}
\alias{make_ui_NOMC}
\title{Make NOMC checkboxes}
\usage{
make_ui_NOMC(
  label,
  choiceNames,
  choiceValues,
  sublabel,
  labels = NULL,
  hide = FALSE,
  arrange_vertically = length(choiceNames) > 2L,
  id = "response_ui"
)
}
\arguments{
\item{labels}{Optional vector of labels for the NOMC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{hide}{Whether the checkboxes should be hidden
(possibly to be shown later).}

\item{arrange_vertically}{Whether to arrange the checkboxes vertically
(the default) as opposed to horizontally.}

\item{id}{HTML ID for the div containing the checkboxes.}

\item{choices}{Character vector of choices for the participant.
If unnamed, then these values will be used both for checkboxes IDs
and for checkboxes labels.
If named, then values will be used for checkboxes IDs and names
will be used for checkboxes labels.}
}
\description{
Creates HTML code for n-option multiple choice response options.
}
