% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MLP_standalone}
\alias{MLP_standalone}
\title{MLP Standalone}
\usage{
MLP_standalone(languages = psyquest::languages(), ...)
}
\arguments{
\item{languages}{(Character vector) Verfügbare Sprachen, Standard ist die erste.}

\item{...}{Weitere Argumente an \code{\link{standalone}()}.}
}
\description{
Startet eine eigenständige Testsitzung für das Musical Learning Protocol.
}
