% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{month_and_year_select_page.validate}
\alias{month_and_year_select_page.validate}
\title{Validate month and year from the month and year select page}
\usage{
month_and_year_select_page.validate(show_month = T)
}
\description{
Validate month and year from the month and year select page
}
