% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SCA_standalone}
\alias{SCA_standalone}
\title{SCA Standalone
This function launches a standalone testing session for the SCA questionnaire.}
\usage{
SCA_standalone(languages = SCA_languages(), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{SCA_standalone}()}.}
}
\description{
SCA Standalone
This function launches a standalone testing session for the SCA questionnaire.
}
