% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SCA_standalone}
\alias{SCA_standalone}
\title{SCA Standalone}
\usage{
SCA_standalone(languages = psyquest::languages(), short_version = FALSE, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the SCA questionnaire.
SCA stands for 'Academic Self-Concept'.
}
