% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDQ.R
\name{SDQ}
\alias{SDQ}
\title{SDQ}
\usage{
SDQ(label = "SDQ", dict = psyquest::psyquest_dict)
}
\arguments{
\item{label}{(Character scalar) Label to give the SDQ results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}
}
\description{
This function defines a SDQ module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SDQ in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the SDQ, consider using \code{\link{SDQ_demo}()}.
For a standalone implementation of the SDQ,
consider using \code{\link{SDQ_standalone}()}.
}
