% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SDQ_standalone}
\alias{SDQ_standalone}
\title{SDQ Standalone
This function launches a standalone testing session for the SDQ questionnaire.}
\usage{
SDQ_standalone(languages = SDQ_languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}), and German (\code{"de"}).
The first language is selected by default.}

\item{subscales}{(Character vector)
Determines the subscales to be included.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{SDQ_standalone}()}.}
}
\description{
SDQ Standalone
This function launches a standalone testing session for the SDQ questionnaire.
}
