% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SDQ_standalone}
\alias{SDQ_standalone}
\title{SDQ Standalone}
\usage{
SDQ_standalone(languages = psyquest::languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"en"} (English), and \code{"de"} (German).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Conduct problems"}, \code{"Difficulties"}, \code{"Emotional problems"}, \code{"Externalising"}, \code{"Hyperactivity"}, \code{"Internalising"}, \code{"Peer problems"}, and \code{"Prosocial"}.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for the SDQ questionnaire.
SDQ stands for 'Strengths and Difficulties Questionnaire (mental health)'.
}
