% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM}
\alias{SEM}
\title{SEM}
\usage{
SEM(label = "SEM", dict = psyquest::psyquest_dict)
}
\arguments{
\item{label}{(Character scalar) Label to give the SEM results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}
}
\description{
This function defines a SEM module for incorporation into a
psychTestR timeline.
Use this function if you want to include the SEM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the SEM, consider using \code{\link{SEM_demo}()}.
For a standalone implementation of the SEM,
consider using \code{\link{SEM_standalone}()}.
}
