% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SEM_standalone}
\alias{SEM_standalone}
\title{SEM Standalone
This function launches a standalone testing session for the SEM questionnaire.}
\usage{
SEM_standalone(languages = c("EN", "DE"), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}
}
\description{
SEM Standalone
This function launches a standalone testing session for the SEM questionnaire.
}
