% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SES_standalone}
\alias{SES_standalone}
\title{SES Standalone
This function launches a standalone testing session for the SES questionnaire.}
\usage{
SES_standalone(languages = SES_languages(), subscales = NULL, ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}), and German (\code{"de"}).
The first language is selected by default.}

\item{subscales}{(Character vector)
Determines the subscales to be included.
If no subscales are provided all subscales for the questionnaire are selected.}

\item{...}{Further arguments to be passed to \code{\link{SES_standalone}()}.}
}
\description{
SES Standalone
This function launches a standalone testing session for the SES questionnaire.
}
