% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP.R
\name{SMP}
\alias{SMP}
\title{SMP(STOMP)}
\usage{
SMP(label = "SMP", dict = psyquest::psyquest_dict, subscales = c(), ...)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are \code{"Reflective & Complex"}, \code{"Intense & Rebellious"},
\code{"Upbeat & Conventional"}, and \code{"Energetic & Rhythmic"}.
If no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{SMP}()}.}
}
\description{
This function defines a SMP module for incorporation into a
psychTestR timeline.
Use this function if you want to include the STOMP in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the SMP,
consider using \code{\link{SMP_standalone}()}.
}
