% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone}
\alias{standalone}
\title{Standalone}
\usage{
standalone(
  questionnaire = questionnaire,
  admin_password = "conifer",
  researcher_email = "musicsophistication@gmail.com",
  languages = languages,
  dict = psyquest::psyquest_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{questionnaire}{(Scalar character) The questionnaire acronym.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto"
for default validation which means ID should consist only of alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{DAC}()}.}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end
of the test. Defaults to FALSE.}

\item{take_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}
}
\description{
This function launches a standalone testing session for a given questionnaire.
This can be used for data collection, either in the laboDACory or online.
}
