% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone}
\alias{standalone}
\title{Standalone}
\usage{
standalone(
  label,
  languages = c("EN", "DE"),
  subscales = NULL,
  short_version = FALSE,
  randomise = FALSE,
  configuration_filepath = NULL,
  dict = psyquest::psyquest_dict,
  admin_password = "conifer",
  researcher_email = "musicsophistication@gmail.com",
  validate_id = "auto",
  with_feedback = FALSE,
  take_training = TRUE,
  ...
)
}
\arguments{
\item{label}{(Scalar character) Three uppercase letter acronym of the questionnaire.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
When no subscales are provided all subscales are selected.}

\item{short_version}{(Scalar boolean) For the short version of the questionnaire set this to TRUE.
Defaults to FALSE.}

\item{randomise}{(Boolean scalar) Defines whether questionnaire items appear in
randomised order. Defaults to FALSE.}

\item{configuration_filepath}{(Character scalar) Optional path to a configuration file exported from the GMSI-Configurator at https://shiny.gold-msi.org/gmsiconfigurator (GMS only).}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto"
for default validation which means ID should consist only of alphanumeric characters.}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end
of the test. Defaults to FALSE.}

\item{take_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for a given questionnaire.
This can be used for data collection, either in the laboratory or online.
}
