% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOI.R
\name{TOI}
\alias{TOI}
\title{TOI}
\usage{
TOI(
  label = "TOI",
  dict = psyquest::psyquest_dict,
  items = items,
  subscales = subscales,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label to give the TOI results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{items}{(Data frame) The items to be included in the questionnaire.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
When no subscales are provided all subscales are selected.}

\item{...}{Further arguments to be passed to \code{\link{TOI}()}.}
}
\description{
This function defines a TOI module for incorporation into a
psychTestR timeline.
Use this function if you want to include the TOI in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the TOI,
consider using \code{\link{TOI_standalone}()}.
}
