% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOM.R
\name{TOM}
\alias{TOM}
\title{TOM}
\usage{
TOM(label = "TOM", dict = psyquest::psyquest_dict, ...)
}
\arguments{
\item{label}{(Character scalar) Label to give the TOM results in the output file.}

\item{dict}{The psyquest dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{TOM}()}.}
}
\description{
This function defines a TOM module for incorporation into a
psychTestR timeline.
Use this function if you want to include the TOM in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the TOM,
consider using \code{\link{TOM_standalone}()}.
}
