% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{TPI_standalone}
\alias{TPI_standalone}
\title{TPI Standalone
This function launches a standalone testing session for the TPI questionnaire.}
\usage{
TPI_standalone(languages = c("EN", "DE"), ...)
}
\arguments{
\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}
}
\description{
TPI Standalone
This function launches a standalone testing session for the TPI questionnaire.
}
